+++ frontend/src/proto/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export * from "./envelope";
+export * from "./messages";
+export * from "./errors";